// src/app/shared/enums/users-type.action.ts
export const enum usersTypeActions {

  LOAD_USERS = '[User] - Loading users',
  LOAD_USERS_ERROR = '[User] - Load users (e)rror!',
  LOAD_USERS_SUCCESS = '[User] - Load users sucess!',

  LOAD_USER = '[User] - Loading user',
  LOAD_USER_ERROR = '[User] - Load users (e)rror!',
  LOAD_USER_SUCCESS = '[User] - Load user sucess!',

  ADD_USER = '[User] - Add user',
  ADD_USER_ERROR = '[User] - Add users (e)rror!',
  ADD_USER_SUCCESS = '[User] - Add users success!',

  DELETE_USER = '[User] - Delete user',
  DELETE_USER_ERROR = '[User] - Delete users (e)rror!',
  DELETE_USER_SUCCESS =  '[User] - Delete users success!',

  UPDATE_USER = '[User] - Update user',
  UPDATE_USER_ERROR = '[User] - Update users (e)rror!',
  UPDATE_USER_SUCCESS = '[User] - Update users success!',

}